Move obtain-access-token.js to /authorise directory
