+++ app/authorise/authorisation-code-granted.js
@@ -1,6 +1,6 @@
 const env = require('env-var');
 const { setTokenPayload } = require('./access-tokens');
-const { postToken } = require('../obtain-access-token');
+const { postToken } = require('./obtain-access-token');
 const { getClientCredentials } = require('../authorisation-servers');
 const debug = require('debug')('debug');
